+++ admin-service/src/main/resources/pipeline-script-samples/csv.groovy
@@ -1,1 +1,7 @@
-// csv sample script goes here
+// enables csv utilities
+enable 'csv'
+
+// add your transformations here
+
+// outputs transformed csv
+outputCsv