+++ webapp/src/main/webapp/src/app/shared/i18n/rdw-translate-loader.ts
@@ -1,23 +1,32 @@
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 import { merge, set, get } from 'lodash';
-import { Injectable } from '@angular/core';
+import { Injectable, Optional } from '@angular/core';
 import { TranslateLoader } from '@ngx-translate/core';
-import { Observable, forkJoin, of } from 'rxjs';
+import { Observable, forkJoin, of, iif } from 'rxjs';
 import { EmbeddedLanguages } from './language-settings';
 import { HttpClient } from '@angular/common/http';
-import { catchError, map } from 'rxjs/operators';
+import { catchError, map, mergeMap } from 'rxjs/operators';
 import { SubjectService } from '../../subject/subject.service';
 import { flattenJsonObject, Utils } from '../support/support';
+import { UserService } from '../../user/user.service';
 
 const EmptyObservable = of({});
 const AssessmentTypes: string[] = ['iab', 'ica', 'sum'];
+const defaultUserService = <UserService>{
+  getUser: () => of({ anonymous: true })
+};
 
 @Injectable()
 export class RdwTranslateLoader implements TranslateLoader {
   private clientTranslationsLoader;
   private serverTranslationsLoader;
 
-  constructor(http: HttpClient, private subjectService: SubjectService) {
+  constructor(
+    http: HttpClient,
+    private subjectService: SubjectService,
+    @Optional()
+    private userService: UserService = defaultUserService
+  ) {
     this.clientTranslationsLoader = new TranslateHttpLoader(
       http,
       '/assets/i18n/',
@@ -31,19 +40,15 @@ export class RdwTranslateLoader implements TranslateLoader {
   }
 
   getTranslation(languageCode: string): Observable<any> {
-    return forkJoin(
-      this.getClientTranslations(languageCode),
-      this.getServerTranslations(languageCode),
-      this.subjectService.getSubjectCodes()
-    ).pipe(
-      map(([clientTranslations, serverTranslations, subjects]) => {
-        const asmtTranslations = this.createSubjectTranslations(
-          subjects,
-          serverTranslations
-        );
-        return merge(clientTranslations, asmtTranslations, serverTranslations);
-      })
-    );
+    return this.userService
+      .getUser()
+      .pipe(
+        mergeMap(user =>
+          user.anonymous
+            ? this.getClientTranslations(languageCode)
+            : this.getUserTranslations(languageCode)
+        )
+      );
   }
 
   getFlattenedTranslations(languageCode: string): Observable<any> {
@@ -112,5 +117,21 @@ export class RdwTranslateLoader implements TranslateLoader {
     }
 
     return subjectTranslations;
+  }
+
+  private getUserTranslations(languageCode: string) {
+    return forkJoin(
+      this.getClientTranslations(languageCode),
+      this.getServerTranslations(languageCode),
+      this.subjectService.getSubjectCodes()
+    ).pipe(
+      map(([clientTranslations, serverTranslations, subjects]) => {
+        const asmtTranslations = this.createSubjectTranslations(
+          subjects,
+          serverTranslations
+        );
+        return merge(clientTranslations, asmtTranslations, serverTranslations);
+      })
+    );
   }
 }