+++ webapp/src/main/webapp/src/app/user/user.ts
@@ -2,4 +2,5 @@ export interface User {
   readonly firstName: string;
   readonly lastName: string;
   readonly permissions: string[];
+  readonly anonymous?: boolean;
 }